# docs/swagger.yaml
definitions:
  errorHandler.HttpErr:
    properties:
      message:
        example: Error message
        type: string
      status_code:
        example: 400
        type: integer
    type: object
...
  input.UserLogin:
    properties:
      email:
        example: test@gmail.com
        type: string
      password:
        example: "123456"
        type: string
    type: object
  input.UserRegister:
    properties:
      email:
        example: test@gmail.com
        type: string
      first_name:
        example: Иван
        type: string
      last_name:
        example: Иванов
        type: string
      password:
        example: "123456"
        type: string
      patronymic:
        example: Иванович
        type: string
      phone:
        example: "89121234567"
        type: string
    type: object
...
  response.AppealCategory:
    properties:
      id:
        example: 1
        type: integer
      title:
        example: Категория
        type: string
    type: object
...
  response.AppealStatus:
    properties:
      id:
        example: 1
        type: integer
      status:
        example: Решено
        type: string
    type: object
  response.AppealType:
    properties:
      appeal_category:
        $ref: '#/definitions/response.AppealCategory'
      id:
        example: 1
        type: integer
      title:
        example: Тип
        type: string
    type: object
...
  response.MapElement:
    properties:
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      title:
        type: string
      type:
        type: string
    type: object
  response.News:
    properties:
      address:
        example: Улица Пушкина
        type: string
      body:
        example: Тело новости
        type: string
      category:
        $ref: '#/definitions/response.NewsCategory'
      date:
        example: "2024-02-10T00:00:00+05:00"
        type: string
      id:
        example: 1
        type: integer
      latitude:
        example: 54.1234
        type: number
      longitude:
        example: 122.7656
        type: number
      photo_url:
        example: https://storage.yandexcloud.net/urbathon/test.jpg
        type: string
      poll:
        $ref: '#/definitions/response.NewsPoll'
      title:
        example: Заголовок
        type: string
    type: object
  response.NewsCategory:
    properties:
      id:
        example: 1
        type: integer
      title:
        example: Отходы
        type: string
    type: object
  response.NewsPaged:
    properties:
      items:
        items:
          $ref: '#/definitions/response.News'
        type: array
      next_page_url:
        example: /api/news?page=1&page_size=10
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
  response.NewsPoll:
    properties:
      id:
        example: 1
        type: integer
      options:
        items:
          $ref: '#/definitions/response.PollOption'
        type: array
      title:
        example: Название опроса
        type: string
    type: object
  response.Organization:
    properties:
      address:
        type: string
      id:
        example: 1
        type: integer
      inn:
        type: string
      name:
        example: ЖКХ
        type: string
      phone:
        type: string
    type: object
  response.PollOption:
    properties:
      id:
        example: 1
        type: integer
      is_user_voted:
        type: boolean
      title:
        example: Вариант опроса
        type: string
      votes:
        example: 1
        type: integer
    type: object
  response.User:
    properties:
      email:
        example: test@gmail.com
        type: string
      first_name:
        example: Иван
        type: string
      id:
        example: 1
        type: integer
      job:
        type: string
      last_name:
        example: Иванов
        type: string
      organization:
        $ref: '#/definitions/response.Organization'
      organization_id:
        example: 1
        type: integer
      patronymic:
        example: Иванович
        type: string
    type: object
info:
  contact: {}
  description: Спецификация приложения команды подCRUDули
  title: Urbathon-2023
paths:
...
  /appeal_status:
    get:
      description: Получить все статусы обращений
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.AppealStatus'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получить все статусы обращений
      tags:
      - appealStatus
  /appeal_status/{id}:
    get:
      description: Получить статус обращения
      parameters:
      - default: 1
        description: appealStatus id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AppealStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получить статус обращения
      tags:
      - appealStatus
  /appeal_type:
    get:
      description: Получить все типы обращений
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.AppealType'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получить все типы обращений
      tags:
      - appealType
  /appeal_type/{id}:
    get:
      description: Получить тип обращения по айди
      parameters:
      - description: appealType id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AppealType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получить тип обращения по айди
      tags:
      - appealType
  /auth/login:
    post:
      consumes:
      - application/json
      description: Вход
      parameters:
      - description: login and password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/input.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Вход
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрация
      parameters:
      - description: UserRegister
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/input.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Регистрация
      tags:
      - auth
...
  /map/get_map_elements:
    get:
      description: Получение всех маркеров на карте
      parameters:
      - in: query
        name: lat_down
        type: number
      - in: query
        name: lat_up
        type: number
      - in: query
        name: long_down
        type: number
      - in: query
        name: long_up
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.MapElement'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получение всех маркеров на карте
      tags:
      - map
  /news:
    get:
      description: Получение всех новостей
      parameters:
      - default: 1
        description: page
        in: query
        minimum: -1
        name: page
        type: integer
      - default: 10
        description: page
        in: query
        maximum: 20
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NewsPaged'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получение всех новостей
      tags:
      - news
    post:
      consumes:
      - multipart/form-data
      description: Создание новости
      parameters:
      - example: Улица Пушкина
        in: formData
        name: address
        type: string
      - in: formData
        name: body
        type: string
      - in: formData
        name: category_id
        type: integer
      - example: 54.1234
        in: formData
        name: latitude
        type: number
      - example: 122.7656
        in: formData
        name: longitude
        type: number
      - example: тут json голосования
        in: formData
        name: poll
        type: string
      - in: formData
        name: title
        type: string
      - description: photo
        in: formData
        name: photo
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Создание новости
      tags:
      - news
  /news/{id}:
    get:
      description: Получение новости по айди
      parameters:
      - description: news id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Получение новости по айди
      tags:
      - news
  /news/{id}/poll_vote/{option_id}:
    post:
      description: Проголосовать в опросе под новостью
      parameters:
      - description: news id
        in: path
        name: id
        required: true
        type: integer
      - description: option_id
        in: path
        name: option_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.News'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorHandler.HttpErr'
      summary: Проголосовать в опросе под новостью
      tags:
      - news
...
swagger: "2.0"